refactor: 스토리지 상수 위치 이동 및 초대 코드 상수 네이밍 수정
