+++ src/constants/api/index.ts
@@ -3,4 +3,3 @@ export * from './endpoints';
 export * from './httpStatus';
 export * from './errorCode';
 export * from './errorMessage';
-export * from './storage';
