--- src/constants/api/storage.ts
@@ -1,3 +1,0 @@
-export const ACCESS_TOKEN = 'ACCESS_TOKEN';
-
-export const INVITE_URL = 'INVITE_URL';
