+++ src/constants/storage.ts
@@ -1,0 +1,3 @@
+export const ACCESS_TOKEN = 'ACCESS_TOKEN';
+
+export const INVITE_URL_KEY = 'INVITE_URL_KEY';
