+++ src/hooks/queries/useLoginMutation.tsx
@@ -4,12 +4,13 @@ import { Stomp } from '@stomp/stompjs';
 import { useMutation } from '@tanstack/react-query';
 import SockJS from 'sockjs-client';
 import useSocketStore from '@stores/socketStore';
-import { ACCESS_TOKEN, INVITE_URL, WEBSOCKET_URL } from '@constants/api';
+import { WEBSOCKET_URL } from '@constants/api';
 import { PATH } from '@constants/path';
+import { ACCESS_TOKEN, INVITE_URL_KEY } from '@constants/storage';
 
 const useLoginMutation = () => {
 	const navigate = useNavigate();
-	const inviteUrl = window.sessionStorage.getItem(INVITE_URL);
+	const inviteUrl = window.sessionStorage.getItem(INVITE_URL_KEY);
 	const { setStompClient } = useSocketStore();
 
 	const postLoginMutation = useMutation({
@@ -18,9 +19,7 @@ const useLoginMutation = () => {
 			localStorage.setItem(ACCESS_TOKEN, content.accessToken);
 
 			const client = Stomp.over(function () {
-				return new SockJS(
-					`${WEBSOCKET_URL}${localStorage.getItem(ACCESS_TOKEN)}`
-				);
+				return new SockJS(`${WEBSOCKET_URL}${localStorage.getItem(ACCESS_TOKEN)}`);
 			});
 
 			client.connect({}, () => {
@@ -30,7 +29,7 @@ const useLoginMutation = () => {
 			client.debug = () => {};
 
 			if (inviteUrl) {
-				window.sessionStorage.removeItem(INVITE_URL);
+				window.sessionStorage.removeItem(INVITE_URL_KEY);
 				navigate(`${PATH.INVITE}${inviteUrl}`, { replace: true });
 			} else if (content.hasTeam) {
 				navigate(PATH.FEED, { replace: true });
