+++ src/pages/LandingPage/LandingPage.tsx
@@ -2,8 +2,8 @@ import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Button } from '@components/common/Button/Button';
 import Text from '@components/common/Text/Text';
-import { ACCESS_TOKEN } from '@constants/api';
 import { PATH } from '@constants/path';
+import { ACCESS_TOKEN } from '@constants/storage';
 import { collaBear } from '@assets/png';
 import { Colla } from '@assets/svg';
 import * as S from './LandingPage.styled';
